+++ src/Form/SettingsForm.php
@@ -83,6 +83,24 @@ class SettingsForm extends ConfigFormbase {
       '#default_value' => $config->get('intake_email'),
     ];
 
+    // RCP template.
+    $form['rcp_template_path'] = [
+      '#type' => 'textfield',
+      '#title' => $this->t('Resource conservation plan template path'),
+      '#description' => $this->t('Leave blank to use the default template.'),
+      '#default_value' => $config->get('rcp_template_path'),
+    ];
+    $form['rcp_template_upload'] = [
+      '#type' => 'file',
+      '#title' => $this->t('Upload resource conservation plan template'),
+      '#description' => $this->t('Must be a Word 2007 compatible document file with a ".docx" extension.'),
+      '#upload_validators' => [
+        'FileExtension' => [
+          'extensions' => 'docx',
+        ],
+      ],
+    ];
+
     return parent::buildForm($form, $form_state);
   }
 
@@ -94,9 +112,9 @@ class SettingsForm extends ConfigFormbase {
 
     // Check for a new uploaded logo and put the temporary file in form values,
     // so we can save it on submit.
-    $file = _file_save_upload_from_form($form['logo_upload'], $form_state, 0);
-    if ($file) {
-      $form_state->setValue('logo_upload', $file);
+    $logo_file = _file_save_upload_from_form($form['logo_upload'], $form_state, 0);
+    if ($logo_file) {
+      $form_state->setValue('logo_upload', $logo_file);
     }
 
     // If the user provided a path for a logo file, make sure a file exists
@@ -105,6 +123,22 @@ class SettingsForm extends ConfigFormbase {
       $path = $this->validatePath($form_state->getValue('logo_path'));
       if (!$path) {
         $form_state->setErrorByName('logo_path', $this->t('The logo path is invalid.'));
+      }
+    }
+
+    // Check for a new uploaded RCP template and put the temporary file in form
+    // values, so we can save it on submit.
+    $rcp_template_file = _file_save_upload_from_form($form['rcp_template_upload'], $form_state, 0);
+    if ($rcp_template_file) {
+      $form_state->setValue('rcp_template_upload', $rcp_template_file);
+    }
+
+    // If the user provided a path for an RCP template file, make sure a file
+    // exists at that path.
+    if ($form_state->getValue('rcp_template_path')) {
+      $path = $this->validatePath($form_state->getValue('rcp_template_path'));
+      if (!$path) {
+        $form_state->setErrorByName('rcp_template_path', $this->t('The resource conservation plan template path is invalid.'));
       }
     }
   }
@@ -150,28 +184,39 @@ class SettingsForm extends ConfigFormbase {
   public function submitForm(array &$form, FormStateInterface $form_state) {
     $values = $form_state->getValues();
 
-    // If the user uploaded a new logo, save it to a permanent location.
+    // If the user uploaded a new logo or RCP template, save them to a
+    // permanent location.
     try {
       if (!empty($values['logo_upload'])) {
         $filename = $this->fileSystem->copy($values['logo_upload']->getFileUri(), 'public://');
         $values['logo_path'] = $filename;
+      }
+      if (!empty($values['rcp_template_upload'])) {
+        $filename = $this->fileSystem->copy($values['rcp_template_upload']->getFileUri(), 'public://');
+        $values['rcp_template_path'] = $filename;
       }
     }
     catch (FileException) {
       // Ignore.
     }
     unset($values['logo_upload']);
+    unset($values['rcp_template_upload']);
 
     // If the user entered a path relative to the system files directory for a
-    // logo, store a public:// URI so the theme system can handle it.
+    // logo or RCP template, store a public:// URI so the theme system can
+    // handle it.
     if (!empty($values['logo_path'])) {
       $values['logo_path'] = $this->validatePath($values['logo_path']);
+    }
+    if (!empty($values['rcp_template_path'])) {
+      $values['rcp_template_path'] = $this->validatePath($values['rcp_template_path']);
     }
 
     // Save to config.
     $this->configFactory->getEditable(static::SETTINGS)
       ->set('logo_path', $values['logo_path'])
       ->set('intake_email', $values['intake_email'])
+      ->set('rcp_template_path', $values['rcp_template_path'])
       ->save();
 
     parent::submitForm($form, $form_state);